refactor(user-ui): update popup service
