+++ packages/user-ui/src/components/mod.rs
@@ -4,9 +4,11 @@ pub mod header;
 pub mod icons;
 pub mod input;
 pub mod label;
+pub mod popup_zone;
 pub mod search;
 pub mod textarea;
+
 pub use footer::Footer;
 pub use header::Header;
-
 pub mod button;
+pub use popup_zone::PopupZone;
