+++ packages/user-ui/src/components/popup_zone/mod.rs
@@ -1,81 +1,5 @@
-#![allow(non_snake_case)]
-use dioxus::prelude::*;
-
-#[derive(Debug, Clone, Copy, Default)]
-pub struct PopupService {
-    pub id: Signal<Option<String>>,
-    pub title: Signal<Option<String>>,
-    pub data: Signal<Option<Element>>,
-    pub close: Signal<bool>,
-}
-
-impl PopupService {
-    pub fn init() {
-        let srv = Self {
-            data: Signal::new(None),
-            id: Signal::new(None),
-            title: Signal::new(None),
-            close: Signal::new(true),
-        };
-        use_context_provider(|| srv);
-    }
-
-    pub fn render(&self) -> Element {
-        (self.data)().clone().unwrap_or(default())
-    }
-
-    pub fn is_opened(&self) -> bool {
-        (self.data)().is_some()
-    }
-
-    pub fn get_id(&self) -> String {
-        (self.id)().clone().unwrap_or("popup-zone".to_string())
-    }
-
-    pub fn get_title(&self) -> Option<String> {
-        (self.title)().clone()
-    }
-
-    pub fn open(&mut self, popup: Element) -> &mut Self {
-        (self.data).set(Some(popup));
-
-        self
-    }
-
-    pub fn with_id(&mut self, id: &str) -> &mut Self {
-        (self.id).set(Some(id.to_string()));
-
-        self
-    }
-
-    pub fn with_title(&mut self, title: &str) -> &mut Self {
-        (self.title).set(Some(title.to_string()));
-
-        self
-    }
-
-    pub fn without_close(&mut self) -> &mut Self {
-        (self.close).set(false);
-
-        self
-    }
-
-    pub fn close(&mut self) {
-        (self.data).set(None);
-        (self.id).set(None);
-        (self.title).set(None);
-        (self.close).set(true);
-    }
-
-    pub fn use_popup_service() -> PopupService {
-        use_context()
-    }
-}
-
-#[component]
-pub fn default() -> Element {
-    rsx! {}
-}
+use bdk::prelude::*;
+use dioxus_popup::PopupService;
 
 #[component]
 pub fn PopupZone() -> Element {
