+++ packages/user-ui/src/main.rs
@@ -1,11 +1,10 @@
 use bdk::prelude::*;
-
+use dioxus_popup::PopupService;
 pub mod components;
 pub mod config;
 pub mod pages;
 
 pub mod service {
-    pub mod popup_service;
     pub mod user_service;
 }
 
@@ -17,7 +16,7 @@ pub mod routes;
 
 use dioxus_oauth::prelude::FirebaseProvider;
 use routes::Route;
-use service::{popup_service::PopupService, user_service::UserService};
+use service::user_service::UserService;
 
 const FAVICON: Asset = asset!("/public/favicon.png");
 const MAIN_CSS: Asset = asset!("/public/main.css");
